added saveStateCapabilities

+++ include/gameUpgrades.h
@@ -9,7 +9,7 @@ TODO: put this in an extra file
 
 GameUpgrade Preset Specification
 Must have 255 lines
-first line denotes the UpgradeCost(int), unitPerCycle(int) and framesPerCycle(float), delimited by space, at level 0;
+first line denotes the UpgradeCost(int), unitPerCycle(int) and framesPerCycle(int), delimited by space, at level 0;
 the n-th line represents the change in the aforemention stat at n-th level
  */
